Removed disable trigger interceptor and moved trigger disabling into output-inserted insert handler (as they are tightly coupled) and removed output-inserted insert handler so it is not part of default configuration
